Look in current directory for assembly and Remove check for trailing \
Previous application logic looked in the path where the assembly was
currently executing from. In Mono this is a temporary directory that
only contains the assembly not the configuration information. Using the
GetCurrentDirectory() call returns a path that has the configuration
directory.

The check for hard coded "\\" is not cross platform compatible. Replaced
with Path.DirectorySeparatorChar.